requirements.txt
evidently==0.4.36
mlflow==2.15.1
databricks-feature-store==0.17.0
pyspark==3.2.0
black==24.8.0
isort==5.13.2
pytest==8.3.2 